# [xamlg] improve performance by reusing Mono.Cecil objects