### Description of Change ###

I found that for each iteration of a `XamlFile`, `<XamlGTask/>` would:

* Iterate over each item in `@(ReferencePath)`.
* Load the assembly with Mono.Cecil.
* Find all the `XmlnsDefinitionAttribute` and create dictionaries.

This would happen for every `.xaml.g.cs` file generated. In
`Xamarin.Forms.Controls.csproj` this resulted in 122 references X 379
`.xaml.g.cs` files.

I moved this logic into a new `XmlnsCache` class. This work can be
done once, and `_xmlnsDefinitions` and `_xmlnsModules` can be reused
for each `.xaml.g.cs` file.

This changes the "lifetime" of `XamlGenerator` a bit, now unit tests
should dispose of `XamlGenerator`, while the `<XamlGTask/>` should
only dispose the `XmlnsCache` after all work is complete.

I also wrote the `XmlnsCache`, so the assemblies are loaded lazily if
the `XmlnsDefinitions` or `XmlnsModules` properties are not accessed.

Other small improvements:

* Don't create the `$(ReferencedAssemblies)` property. This is
  equivalent to doing a `string.Join()` with many full paths.
* We can use `@(ReferencePath)` as-is, to pass into a `string[]`
  property to the `<XamlGTask/>`. This way we don't have to split on
  `;` ourselves.
* The `AssemblyIsSystem()` method should use
  `StringComparison.OrdinalIgnoreCase` as `CurrentCultureIgnoreCase`
  is slower and can produce incorrect results in some cultures.

### Results ###

On Windows, when building `Xamarin.Forms.Controls`:

    Before:
    2643 ms  XamlG                                      1 calls
    2626 ms  XamlGTask                                  1 calls
    After:
     388 ms  XamlG                                      1 calls
     374 ms  XamlGTask                                  1 calls

On macOS, running under Mono:

    Before:
    5003 ms  XamlG                                      1 calls
    4926 ms  XamlGTask                                  1 calls
    After:
     772 ms  XamlG                                      1 calls
     688 ms  XamlGTask                                  1 calls

`Xamarin.Forms.Controls` might resemble a larger customer project?
These changes would help the most for projects with lots of
`.xaml.g.cs` files.

### Issues Resolved ### 

None

### API Changes ###
 
None

### Platforms Affected ### 

- Core/XAML (all platforms)

### Behavioral/Visual Changes ###

None

### Before/After Screenshots ### 

Not applicable

### Testing Procedure ###

We may want to do some manual testing in IDEs, I can help with this.

### PR Checklist ###

- [ ] Targets the correct branch
- [ ] Tests are passing (or failures are unrelated)