@StephaneDelcroix 2021-01-04
Thanks for investigating this, and proposing this patch. I love it.

It's ok to keep the references alive in a cache, as they are shared across all xaml files of a single assembly. Does this make sure the cache is invalidated when XamlG runs on xaml files in multiples assemblies ?