@jonathanpeppers 2021-01-04
> Does this make sure the cache is invalidated when XamlG runs on xaml files in multiples assemblies ?

Every time the `<XamlGTask/>` runs it creates a new `XmlnsCache` that is disposed at the end of `Execute()`. So there should be a fresh instance of `XmlnsCache` for each project/assembly.