@transis2 2021-01-13
This kind of improvement should be done with
All Net 5+ projects